refactor: systems have weak_ptrs

+++ Timer.h
@@ -16,14 +16,14 @@ enum format {
 class Timer {
 public:
 	static float difference(time_point<steady_clock> init, time_point<steady_clock> current = Timer::now(), format f = MILLISECONDS) {
-		float dif = duration_cast<milliseconds>(current - init).count();
+		float dif = static_cast<float>(duration_cast<milliseconds>(current - init).count());
 
 		if (f == MILLISECONDS) return dif;
-		dif /= 1000;
+		dif /= 1000.0f;
 		if (f == SECONDS) return dif;
-		dif /= 60;
+		dif /= 60.0f;
 		if (f == MINUTES) return dif;
-		dif /= 60;
+		dif /= 60.0f;
 
 		return dif;
 	}
